+++ netbox/secrets/tests/test_models.py
@@ -1,3 +1,5 @@
+import string
+
 from Crypto.PublicKey import RSA
 from django.conf import settings
 from django.contrib.auth.models import User
@@ -86,7 +88,7 @@ class SecretTestCase(TestCase):
         """
         Test basic encryption and decryption functionality using a random master key.
         """
-        plaintext = "FooBar123"
+        plaintext = string.printable * 2
         secret_key = generate_random_key()
         s = Secret(plaintext=plaintext)
         s.encrypt(secret_key)