+++ netbox/utilities/filters.py
@@ -1,4 +1,5 @@
 import django_filters
+from taggit.models import Tag
 
 
 class NumericInFilter(django_filters.BaseInFilter, django_filters.NumberFilter):
@@ -19,3 +20,18 @@ class NullableCharFieldFilter(django_filters.CharFilter):
             return super(NullableCharFieldFilter, self).filter(qs, value)
         qs = self.get_method(qs)(**{'{}__isnull'.format(self.name): True})
         return qs.distinct() if self.distinct else qs
+
+
+class TagFilter(django_filters.ModelMultipleChoiceFilter):
+    """
+    Match on one or more assigned tags. If multiple tags are specified (e.g. ?tag=foo&tag=bar), the queryset is filtered
+    to objects matching all tags.
+    """
+    def __init__(self, *args, **kwargs):
+
+        kwargs.setdefault('field_name', 'tags__slug')
+        kwargs.setdefault('to_field_name', 'slug')
+        kwargs.setdefault('conjoined', True)
+        kwargs.setdefault('queryset', Tag.objects.all())
+
+        super(TagFilter, self).__init__(*args, **kwargs)