+++ CHANGELOG.md
@@ -73,6 +73,16 @@ NetBox now supports modeling physical cables for console, power, and interface c
 
 ---
 
+v2.4.8 (FUTURE)
+
+## Bug Fixes
+
+* [#2473](https://github.com/digitalocean/netbox/issues/2473) - Fix encoding of long (>127 character) secrets
+* [#2558](https://github.com/digitalocean/netbox/issues/2558) - Filter on all tags when multiple are passed
+* [#2575](https://github.com/digitalocean/netbox/issues/2575) - Correct model specified for rack roles table
+
+---
+
 v2.4.7 (2018-11-06)
 
 ## Enhancements